feat(fields): read/set all the fields from the given struct
refactor
